Move shared nd-utility functions into _util module
